% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a5.R
\name{a5}
\alias{a5}
\title{A5/1}
\usage{
a5(seed, nonce)
}
\arguments{
\item{seed}{Bit stream used as seed.}

\item{nonce}{Bit stream used as nonce.}
}
\value{
Key stream, to be Xored with plaintext to produce the ciphertext
}
\description{
Already deprecated stream cipher A5/1.
}
