% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes.R
\name{aes}
\alias{aes}
\title{Advanced Encryption Standard. Implemented following the guide
https://nvlpubs.nist.gov/nistpubs/fips/nist.fips.197.pdf}
\usage{
aes(msg, key)
}
\arguments{
\item{msg}{Block of 16 bytes to be encrypted. Represented as a numerical vector
of length 16 whose values go from 0 to 255.}

\item{key}{Key to encrypt the block. Represented as a numerical vector
of length 16/24/32 whose values go from 0 to 255.}
}
\value{
Encrypted block with the same representation as input.
}
\description{
Advanced Encryption Standard. Implemented following the guide
https://nvlpubs.nist.gov/nistpubs/fips/nist.fips.197.pdf
}
\examples{
aes(seq(0,255,17),0:15)
}
