% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{autokey}
\alias{autokey}
\title{Vigenere autokey cipher}
\usage{
autokey(msg, key, decrypt = FALSE, alf_size = 33)
}
\arguments{
\item{msg}{Message.}

\item{key}{Password.}

\item{decrypt}{Encrypt or decrypt mode.}

\item{alf_size}{Alphabet size.}
}
\value{
Encrypted message.
}
\description{
Polyalphabetic, monogramic, aperiodic cipher. Plain text becomes part of the key.
}
\examples{
crypto <- autokey(msg_en,"abc")
crypto
autokey(crypto,"abc",decrypt=TRUE)
}
