% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{beaufort}
\alias{beaufort}
\title{Beaufort cipher}
\usage{
beaufort(msg, key, alf_size = 26)
}
\arguments{
\item{msg}{Message.}

\item{key}{Password.}

\item{alf_size}{Alphabet size.}
}
\value{
Encrypted message.
}
\description{
Polyalphabetic(many ways to substitute letters) monogramic(letter by letter) cipher.
Similar to Vigenere, but encrypt and decrypt are the same function.
}
\examples{
crypto <- beaufort(msg_en,"abc")
crypto
beaufort(crypto,"abc")
}
