% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{caesar}
\alias{caesar}
\title{Caesar cipher}
\usage{
caesar(msg, key, decrypt = FALSE, alf_size = 33)
}
\arguments{
\item{msg}{Message.}

\item{key}{Number of shifts.}

\item{decrypt}{Encrypt or decrypt mode.}

\item{alf_size}{Alphabet size.}
}
\value{
Encrypted message.
}
\description{
Monoalphabetic(only 1 way to substitute letters) monogramic(letter by letter) cipher.
}
\examples{
crypto <- caesar(msg_en,key=3)
crypto
caesar(crypto,key=3,decrypt=TRUE)
}
