% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\name{ddt}
\alias{ddt}
\title{Diferential Distribution Table}
\usage{
ddt(sbox)
}
\arguments{
\item{sbox}{Sbox.}
}
\value{
ddt of the Sbox.
}
\description{
Computes the diferential distribution table of a n·n Sbox
}
\examples{
s <- c(2,3,1,13,4,5,6,7,15,10,8,9,0,12,14,11)
ddt(s)
}
