% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.R
\name{des}
\alias{des}
\title{Data Encryption Standard}
\usage{
des(block, key, decrypt = FALSE)
}
\arguments{
\item{block}{Block to be encrypted. Given as bit stream.}

\item{key}{Key used to encrypt. Given as bit stream.}

\item{decrypt}{Indicates whether it should work in encryption or decryption mode.}
}
\value{
Encrypted block, as bit stream.
}
\description{
This implementation is bugged.
Already deprecated block cipher. Replaced by AES.
}
