% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{freq_analysis}
\alias{freq_analysis}
\title{Frequency analysis}
\usage{
freq_analysis(msg, freq)
}
\arguments{
\item{msg}{Ciphertext.}

\item{freq}{Letter's frequency in the message's language.}
}
\value{
Caesar cipher secret key. Sorted according to their likelihood.
}
\description{
Caesar cipher can be decrypted without the key if the language of the message is known.
}
\examples{
crypto <- cesar(msg_en,3)
crypto
info <- freq_analysis(preprocess(crypto),freqs$en)
info
cesar(crypto,info[1],decrypt=TRUE)
}
