% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{kasiski}
\alias{kasiski}
\title{Kasiski attack}
\usage{
kasiski(msg, freq, lenkey = 0)
}
\arguments{
\item{msg}{Ciphertext.}

\item{freq}{Letter's frequency in the message's language.}

\item{lenkey}{Key length can be provided to avoid compute it.}
}
\value{

}
\description{
Vigenere cipher can be decrypted without the key. First find the key length.
Then divide ciphertext into as many sub-ciphertexts as the key length.
}
\examples{
crypto <- vigenere(msg_en,"abc")
crypto
info <- kasiski(crypto,freqs$en)
info
vigenere(crypto,info$key,decrypt=TRUE)
}
