% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftrot.R
\name{leftrot}
\alias{leftrot}
\title{Rotates the elements of a vector to the left}
\usage{
leftrot(v, rot)
}
\arguments{
\item{v}{Vector to be rotated}

\item{rot}{Number of positions to rotate}
}
\value{
Rotated vector
}
\description{
Rotates the elements of a vector to the left
}
\examples{
leftrot(1:10,2)
leftrot(letters,3)
}
