% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md5.R
\name{md5}
\alias{md5}
\title{Message Digest 5}
\usage{
md5(msg)
}
\arguments{
\item{msg}{Message to digest. Given as a bytestream (numeric array with values from 0 to 255).}
}
\value{
Hashed value as a string.
}
\description{
Message Digest 5
}
\examples{
md5(utf8ToInt(""))
}
