% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc4.R
\name{rc4}
\alias{rc4}
\title{rc4}
\usage{
rc4(msg, key)
}
\arguments{
\item{msg}{Message to encrypt. Given as a bytestream (numeric array with values from 0 to 255).}

\item{key}{Key used for encryption. Given as a bytestream.}
}
\value{
Cipher text as a string.
}
\description{
Already deprecated stream cipher rc4.
}
