% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sha1.R
\name{sha1}
\alias{sha1}
\title{Secure Hash Algorithm 1}
\usage{
sha1(msg)
}
\arguments{
\item{msg}{Message to hash. Given as a bytestream (numeric array with values from 0 to 255).}
}
\value{
Hashed value as a string.
}
\description{
This implementation is bugged.
Already deprecated hash function SHA1. Replaced by SHA3.
}
\examples{
sha1(utf8ToInt(""))
}
