% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toysp.R
\name{toysp}
\alias{toysp}
\title{Toy Substitution-Permutation block cipher}
\usage{
toysp(block, key, rounds, s, p)
}
\arguments{
\item{block}{Block as bit stream.}

\item{key}{Key as bit stream.}

\item{rounds}{Rounds.}

\item{s}{Substitution box n·n, as an array with numbers between 0 and 2^n-1.}

\item{p}{Permuation, as an array with numbers between 1 and block size.}
}
\value{
Encrypted block as bit stream.
}
\description{
Block cipher with substitution permutation and a xor key.
}
\examples{
s <- c(2,3,1,13,4,5,6,7,15,10,8,9,0,12,14,11)
p <- seq(0,5*127,5)\%\%128+1
key <- sample(c(0,1),128,TRUE)
rounds <- 4
block1 <- sample(c(0,1),128,TRUE)
toysp(block1,key,rounds,s,p)
}
