% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classic.R
\name{vigenere}
\alias{vigenere}
\title{Vigenere cipher}
\usage{
vigenere(msg, key, decrypt = FALSE, alf_size = 33)
}
\arguments{
\item{msg}{Message.}

\item{key}{Password.}

\item{decrypt}{Encrypt or decrypt mode.}

\item{alf_size}{Alphabet size.}
}
\value{
Encrypted message.
}
\description{
Polyalphabetic(many ways to substitute letters) monogramic(letter by letter) cipher.
}
\examples{
crypto <- vigenere(msg_en,"abc")
crypto
vigenere(crypto,"abc",decrypt=TRUE)
}
